implement help support and provide basic function implementations
This commit adds two new features:
1. Support for help: This feature allows to automatically generate a
help.  Each model implements 4 functions that provide to the user
information about the internal structure of the model.  This is then
used by the api implementation that call those functions when the client
performs a GET request to the resource url setting the command to SHOW,
ADD, DEL, SET or NONE according to the type of help it wants to get.

2. Basic implementation of some base functions: Some functions can have
a default implementation accross different services, then now the
automatic code generation provides that functions, currently those
functions are
- creation and destruction of an instance
- ports api (create, destroy). Update and get are still to be
  implemented