iovnet_server: don't pass http response object
For the programmer implementing the API it is better to use the
differnet objects representation of the API instead of using directly
the http response object.

TODO: there is still a bug on the list methods, they receive (and not
return) a single instance, while they should return an array.