iovnet_server: fix list paramaters type
The functions operating over a list of elements were using as a
parameter and return case a single element instance, then this commit
changes that in order to use an array of elements.

This commit also removes the use of sharde_ptr in the generated model,as
they are not really necessary and only add complexity, now elements are
saved directly within other elements.